HBSD: Initial proof-of-concept for building vendored Rust applications
This commit introduces bsd.rust.mk, which provides the basic ability to
optionally build base userland applications that have been written in
Rust.

This new addition to the build framework is still very much a
work-in-progress. This commit only serves as a method of saving and
sharing early work.

In order to build Rust components, users must set a new
OPTIONAL_TOOLCHAIN variable to "rust-cargo". Otherwise, Rust components
will be skipped.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>